Lift wxWidgets-specific logic out of CommandManager with new virtuals
